+++ R/withSpinner.R
@@ -18,7 +18,7 @@
 #' Ignored if `image` is used.
 #' @param custom.css Set to `TRUE` if you have your own custom CSS that you defined and you don't want the automatic CSS applied to the spinner.
 #' Ignored if `image` is used.
-#' @param proxy.height If the output UI doesn't specify the output height, you can set a proxy height. It defaults to "400px"
+#' @param proxy.height If the output UI doesn't specify the output height, you can set a proxy height. For example, `textOutput()` usually needs this parameter. It defaults to "400px"
 #' for outputs with undefined height. Ignored if `hide.ui` is set to `FALSE`.
 #' @param id The HTML ID to use for the spinner. If you don't provide one, it will be generated automatically.
 #' @param image The path or URL of the image to use if you want to use a custom image instead of a built-in spinner.
@@ -31,6 +31,9 @@
 #' Setting `hide.ui = FALSE` will result in the spinner showing up on top of the previous output UI.
 #' @param caption Caption to display below the spinner or image (text or HTML). The caption's font color is determined
 #' by the `color` parameter. Ignored if `type` is 1.
+#' @param width The width of the spinner, in pixels. This is only needed in rare cases when the spinner
+#' is not appearing on the screen due to it having no inherent width (for example, when the output is inside
+#' a CSS flexbox without a specified width). Do not use this parameter if the spinner already works.
 #' @seealso [showSpinner()], [hideSpinner()], [showPageSpinner()]
 #' @examples
 #' if (interactive()) {
@@ -64,7 +67,8 @@ withSpinner <- function(
   image.width = getOption("spinner.image.width"),
   image.height = getOption("spinner.image.height"),
   hide.ui = getOption("spinner.hide.ui", default = TRUE),
-  caption = getOption("spinner.caption")
+  caption = getOption("spinner.caption"),
+  width = getOption("spinner.width")
 ) {
 
   if (!inherits(ui_element, "shiny.tag") && !inherits(ui_element, "shiny.tag.list")) {
@@ -85,7 +89,8 @@ withSpinner <- function(
     image.width = image.width,
     image.height = image.height,
     hide.ui = hide.ui,
-    caption = caption
+    caption = caption,
+    width = width
   )
 
   htmltools::attachDependencies(spinner, getDependencies())