+++ inst/assets/spinner.js
@@ -8,6 +8,8 @@ function escapeSelector(s) {
 function show_spinner(id) {
     var selector = "#" + escapeSelector(id);
     var parent = $(selector).closest(".shiny-spinner-output-container");
+    if (parent.length == 0) return;
+
     $(selector).siblings(".load-container, .shiny-spinner-placeholder").removeClass('shiny-spinner-hidden');
 
     if (parent.hasClass("shiny-spinner-hideui")) {
@@ -69,7 +71,7 @@ Shiny.addCustomMessageHandler('shinycssloaders.show_spinner', function(params) {
 });
 
 /* When new value or error comes in, hide spinner container (if any) & show the output */
-$(document).on('shiny:value shiny:error', function(event) {
+$(document).on('shiny:value shiny: error', function(event) {
   var id = event.target.id;
   if (id === undefined) {
     return;